moved ElasticDB firetask to avoid circular import
